+++ src/types/cursor.ts
@@ -1,0 +1,10 @@
+const _cursors = require('../settings/cursors.json')
+
+interface Cursor {
+  name: string,
+  isContinuable?: boolean
+}
+
+const cursors: Cursor[] = _cursors
+
+export { Cursor, cursors }